Move PythonVirtualenvOperator to test_python
